V20210122

+++ services/variantannotation/varank/listener/config/launcher.py
@@ -18,6 +18,7 @@ def create_container_file(containersFile, run, containerName):
 		write_file.write('FOLDER: ' + run + '\n')
 		write_file.write('EXEC_DATE: ' + datetime.now().strftime('%d%m%Y-%H%M%S') + '\n')
 		write_file.write('ID: ' + containerName + '\n')
+	os.chmod(os.path.join(containersFile, containerName + '.log'), 0o777)
 
 
 def create_running_file(run, serviceName):
@@ -73,12 +74,22 @@ def launch(run, serviceName, containersFile, montage, image, launchCommand, conf
 	container_name = serviceName + '_' + os.path.basename(os.path.dirname(run)) + '_' + os.path.basename(run)
 	family_tag = find_family_tag(run)
 	create_container_file(containersFile, run, container_name)
-	if family_tag == '':
-		cmd = 'docker run --rm --name=' + container_name + ' ' + montage + ' ' + image + ' -r ' + run
-		print(cmd)
-		subprocess.call(cmd, shell=True)
-	else:
-		cmd = 'docker run --rm --name=' + container_name + ' ' + montage + ' ' + image + ' -r ' + run + ' -fam ' + family_tag
-		print(cmd)
-		subprocess.call(cmd, shell=True)
+	if os.path.isfile(os.path.join(run, 'POOLComplete.txt')):
+		if family_tag == '':
+			cmd = 'docker run --rm --name=' + container_name + ' ' + montage + ' ' + image + ' -r ' + run + ' -p */POOL/*.final.vcf.gz'
+			print(cmd)
+			subprocess.call(cmd, shell=True)
+		else:
+			cmd = 'docker run --rm --name=' + container_name + ' ' + montage + ' ' + image + ' -r ' + run + ' -fam ' + family_tag + ' -p */POOL/*.final.vcf.gz'
+			print(cmd)
+			subprocess.call(cmd, shell=True)
+	if not os.path.isfile(os.path.join(run, 'POOLComplete.txt')):
+		if family_tag == '':
+			cmd = 'docker run --rm --name=' + container_name + ' ' + montage + ' ' + image + ' -r ' + run
+			print(cmd)
+			subprocess.call(cmd, shell=True)
+		else:
+			cmd = 'docker run --rm --name=' + container_name + ' ' + montage + ' ' + image + ' -r ' + run + ' -fam ' + family_tag
+			print(cmd)
+			subprocess.call(cmd, shell=True)
 	os.umask(original_umask)
